+++ dtslint/spec/data.lint.ts
@@ -1,11 +1,13 @@
 /// <reference types="@types/jest" />
-import { action, Immutable, NgxsDataRepository, StateRepository, Mutable } from '@ngxs-labs/data';
-import { NgxsDataMutablePipe } from '@ngxs-labs/data/utils';
+import { action, NgxsDataRepository, StateRepository } from '@ngxs-labs/data';
 import { NgxsModule, State } from '@ngxs/store';
 import { TestBed } from '@angular/core/testing';
 import { Component, Injectable, Input } from '@angular/core';
+import { Immutable, Mutable } from '@ngxs-labs/data/common';
 import { AsyncPipe } from '@angular/common';
 import { map } from 'rxjs/operators';
+
+import { NgxsDataMutablePipe } from '../../src/utils/external/modules/pipes/ngxs-data-mutable.pipe';
 import { ParentCountModel } from '../../integration/app/examples/count/count.model';
 import { CountSubState } from '../../integration/app/examples/count/count-sub.state';