+++ dtslint/spec/ngxs.lint.ts
@@ -6,15 +6,9 @@ import { Observable, of } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 import { Component, Injectable } from '@angular/core';
 
-import {
-    action,
-    NgxsDataRepository,
-    StateRepository,
-    Persistence,
-    NgxsDataPluginModule,
-    Immutable
-} from '@ngxs-labs/data';
+import { action, NgxsDataRepository, StateRepository, Persistence, NgxsDataPluginModule } from '@ngxs-labs/data';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Immutable } from '@ngxs-labs/data/common';
 
 describe('NGXS Integration', () => {
     it('should be correct patch', () => {