+++ src/interfaces/external.interface.ts
@@ -1,8 +1,8 @@
 import { ActionOptions, ActionType } from '@ngxs/store';
 import { MetaDataModel } from '@ngxs/store/src/internal/internals';
+import { Any, PlainObjectOf } from '@ngxs-labs/data/internals';
+import { Immutable } from '@ngxs-labs/data/common';
 import { Observable } from 'rxjs';
-
-import { Any, PlainObjectOf } from './internal.interface';
 import { Type } from '@angular/core';
 
 /**
@@ -30,23 +30,6 @@ export interface NgxsRepositoryMeta<T = Any> {
     stateMeta?: MetaDataModel;
     operations?: PlainObjectOf<NgxsDataOperation>;
 }
-
-type PrimitiveType = undefined | null | boolean | string | number | Function;
-
-/**
- * @publicApi
- */
-export type ActionEvent = ActionType & { payload: PlainObjectOf<Any> };
-
-/**
- * @publicApi
- */
-export type Immutable<T> = { readonly [K in keyof T]: T[K] extends PrimitiveType ? T[K] : Immutable<T[K]> };
-
-/**
- * @publicApi
- */
-export type Mutable<T> = { -readonly [K in keyof T]: Mutable<T[K]> };
 
 /**
  * @publicApi