+++ src/internals/public-api.ts
@@ -1,0 +1,1 @@
+export { Any, PlainObjectOf } from './types/symbols';