+++ src/repositories/ngxs-data.repository.ts
@@ -5,13 +5,8 @@ import { Observable } from 'rxjs';
 import { action } from '../decorators/action/action';
 import { ngxsDeepFreeze } from '../utils/internals/freeze';
 import { NGXS_DATA_EXCEPTIONS } from '../interfaces/internal.interface';
-import {
-    DataPatchValue,
-    DataRepository,
-    Immutable,
-    ImmutableStateContext,
-    StateValue
-} from '../interfaces/external.interface';
+import { DataPatchValue, DataRepository, ImmutableStateContext, StateValue } from '../interfaces/external.interface';
+import { Immutable } from '../common/types/immutability';
 
 export abstract class NgxsDataRepository<T> implements ImmutableStateContext<T>, DataRepository<T> {
     public readonly name: string;