+++ src/services/ngxs-data-accessor.ts
@@ -1,10 +1,11 @@
 import { NGXS_STATE_CONTEXT_FACTORY, NGXS_STATE_FACTORY, StateClass } from '@ngxs/store/internals';
 import { MappedStore, MetaDataModel } from '@ngxs/store/src/internal/internals';
 import { Inject, Injectable, Injector, NgZone } from '@angular/core';
+import { Any, PlainObjectOf } from '@ngxs-labs/data/internals';
 import { StateContext, Store } from '@ngxs/store';
 
 import { NgxsDataOperation, NgxsRepositoryMeta } from '../interfaces/external.interface';
-import { Any, NGXS_DATA_EXCEPTIONS, PlainObjectOf } from '../interfaces/internal.interface';
+import { NGXS_DATA_EXCEPTIONS } from '../interfaces/internal.interface';
 import { getRepository } from '../utils/internals/ensure-repository';
 
 /**