+++ src/utils/public-api.ts
@@ -1,2 +1,1 @@
 export { NgxsDataUtilsModule } from './external/modules/ngxs-data-utils.module';
-export { NgxsDataMutablePipe } from './external/modules/pipes/ngxs-data-mutable.pipe';