+++ tests/children.spec.ts
@@ -1,8 +1,9 @@
 import { Component, Injectable, OnDestroy, OnInit } from '@angular/core';
 import { Action, NgxsModule, State, StateContext, Store } from '@ngxs/store';
-import { Immutable, NgxsDataPluginModule, NgxsDataRepository, StateRepository } from '@ngxs-labs/data';
+import { NgxsDataPluginModule, NgxsDataRepository, StateRepository } from '@ngxs-labs/data';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { Any, PlainObjectOf } from '../src/interfaces/internal.interface';
+import { Any, PlainObjectOf } from '@ngxs-labs/data/internals';
+import { Immutable } from '@ngxs-labs/data/common';
 
 describe('Check correct deep instance', () => {
     let app: AppState;