+++ tests/registration.spec.ts
@@ -1,9 +1,9 @@
 import { action, NgxsDataPluginModule, NgxsDataRepository, StateRepository } from '@ngxs-labs/data';
 import { NgxsModule, Select, State, Store } from '@ngxs/store';
-import { Any } from '../src/interfaces/internal.interface';
 import { Observable } from 'rxjs';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { Component, OnInit } from '@angular/core';
+import { Any } from '../src/internals/types/symbols';
 
 describe('Check correct deep instance', () => {
     let component: AppComponent;